move pojo gen into separate class
